+++ app/controllers/concerns/index_actions.rb
@@ -1,4 +1,4 @@
-module Listable
+module IndexActions
   ## routed actions
   def list
     render json: create_datatable(serverside: true)