+++ app/datatables/results_datatable.rb
@@ -4,9 +4,9 @@ class ResultsDatatable < IndexDatatable
     cols = [:competition_name, :competition_class, :competition_type, :category, :season,
             :ranking, :skater_name, :nation, :points,
             :short_ranking,
-#            :short_tss, :short_tes, :short_pcs, :short_deductions, :short_bv,
+            :short_tss, :short_tes, :short_pcs, :short_deductions, :short_bv,
             :free_ranking,
-#            :free_tss, :free_tes, :free_pcs, :free_deductions, :free_bv,
+            :free_tss, :free_tes, :free_pcs, :free_deductions, :free_bv,
            ]
 
     super(data, only: cols)