+++ app/models/result.rb
@@ -31,39 +31,9 @@ class Result < ApplicationRecord
     scores.first   ## segment =~ /SHORT/
     #scores.where("scores.segment like ?", 'SHORT%').first
   end
-  def short_tss
-    short.try(:tss)
-  end
-  def short_tes
-    short.try(:tes)
-  end
-  def short_pcs
-    short.try(:pcs)
-  end
-  def short_deductions
-    short.try(:deductions)
-  end
-  def short_bv
-    short.try(:base_value)
-  end
   def free
     #scores.where("scores.segment like ?", 'FREE%').first
     scores.second
-  end
-  def free_tss
-    free.try(:tss)
-  end
-  def free_tes
-    free.try(:tes)
-  end
-  def free_pcs
-    free.try(:pcs)
-  end
-  def free_deductions
-    free.try(:deductions)
-  end
-  def free_bv
-    free.try(:base_value)
   end
   ## scopes
   scope :recent, ->{ joins(:competition).order("competitions.start_date desc") }